formatting by air

+++ R/configMigrate.R
@@ -11,7 +11,6 @@ migrateConfig <- function(configDir) {
 
   # We have no configuration directory but we do have an old one; migrate it.
   if (dirExists(oldConfigDir)) {
-
     # Create the parent folder if necessary
     dirCreate(dirname(configDir))
 
@@ -35,7 +34,6 @@ migrateConfig <- function(configDir) {
 #'
 #' @keywords internal
 oldApplicationConfigDir <- function(appName) {
-
   # get the home directory from the operating system (in case
   # the user has redefined the meaning of ~) but fault back
   # to ~ if there is no HOME variable defined
@@ -51,10 +49,8 @@ oldApplicationConfigDir <- function(appName) {
     # no R specific config dir; determine application config dir (platform specific)
     sysName <- Sys.info()[["sysname"]]
     if (identical(sysName, "Windows"))
-      configDir <- Sys.getenv("APPDATA")
-    else if (identical(sysName, "Darwin"))
-      configDir <- file.path(homeDir, "Library/Application Support")
-    else
+      configDir <- Sys.getenv("APPDATA") else if (identical(sysName, "Darwin"))
+      configDir <- file.path(homeDir, "Library/Application Support") else
       configDir <- Sys.getenv("XDG_CONFIG_HOME", file.path(homeDir, ".config"))
 
     # append the application name
@@ -80,12 +76,17 @@ migrateDeploymentsConfig <- function(appPath) {
   }
 
   migrateDir <- file.path(migrateRoot, "rsconnect")
-  for (shinyappsFile in list.files(shinyappsDir, glob2rx("*.dcf"),
-                                   recursive = TRUE)) {
+  for (shinyappsFile in list.files(
+    shinyappsDir,
+    glob2rx("*.dcf"),
+    recursive = TRUE
+  )) {
     # read deployment record
     shinyappsDCF <- file.path(shinyappsDir, shinyappsFile)
-    deployment <- as.data.frame(read.dcf(shinyappsDCF),
-                                stringsAsFactors = FALSE)
+    deployment <- as.data.frame(
+      read.dcf(shinyappsDCF),
+      stringsAsFactors = FALSE
+    )
     deployment$server <- "shinyapps.io"
 
     # write the new record
@@ -98,9 +99,6 @@ migrateDeploymentsConfig <- function(appPath) {
   }
 
   # remove shinyapps dir if it's completely empty
-  remainingFiles <- list.files(shinyappsDir,
-                               recursive = TRUE,
-                               all.files = TRUE)
-  if (length(remainingFiles) == 0)
-    unlink(shinyappsDir, recursive = TRUE)
+  remainingFiles <- list.files(shinyappsDir, recursive = TRUE, all.files = TRUE)
+  if (length(remainingFiles) == 0) unlink(shinyappsDir, recursive = TRUE)
 }
